+++ pocketsphinx/src/libpocketsphinx/tmat.h
@@ -107,7 +107,7 @@ extern "C" {
  * topology.
  */
 typedef struct {
-    uint8 ***tp;	/**< The transition matrices; kept in the same scale as acoustic scores;
+    int32 ***tp;	/**< The transition matrices; kept in the same scale as acoustic scores;
 			   tp[tmatid][from-state][to-state] */
     int16 n_tmat;	/**< Number matrices */
     int16 n_state;	/**< Number source states in matrix (only the emitting states);
